refactor: centralize pdf fetch logic
